@@ -14,36 +14,35 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^7.1.1",
-    "@angular/cdk": "^7.1.0",
-    "@angular/common": "^7.1.1",
-    "@angular/compiler": "^7.1.1",
-    "@angular/core": "^7.1.1",
-    "@angular/forms": "^7.1.1",
-    "@angular/http": "^7.1.1",
-    "@angular/material": "^7.1.0",
-    "@angular/platform-browser": "^7.1.1",
-    "@angular/platform-browser-dynamic": "^7.1.1",
-    "@angular/router": "^7.1.1",
+    "@angular/animations": "^8.2.2",
+    "@angular/cdk": "^8.1.3",
+    "@angular/common": "^8.2.2",
+    "@angular/compiler": "^8.2.2",
+    "@angular/core": "^8.2.2",
+    "@angular/forms": "^8.2.2",
+    "@angular/material": "^8.1.3",
+    "@angular/platform-browser": "^8.2.2",
+    "@angular/platform-browser-dynamic": "^8.2.2",
+    "@angular/router": "^8.2.2",
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.14.2",
     "ngx-highlightjs": "^3.0.3",
     "ngx-validators": "^4.2.0",
-    "rxjs": "^6.3.3",
-    "zone.js": "~0.8.26"
+    "rxjs": "^6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.1",
-    "@angular-devkit/build-ng-packagr": "~0.11.0",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.1",
-    "@angular/language-service": "^7.1.1",
+    "@angular-devkit/build-angular": "~0.802.2",
+    "@angular-devkit/build-ng-packagr": "~0.802.2",
+    "@angular/cli": "^8.2.2",
+    "@angular/compiler-cli": "^8.2.2",
+    "@angular/language-service": "^8.2.2",
     "@types/googlemaps": "^3.30.16",
     "@types/jasmine": "^2.8.12",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^11.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.0.0",
@@ -51,12 +50,12 @@
     "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.0.0",
+    "ng-packagr": "^5.3.0",
     "protractor": "^5.4.1",
     "ts-node": "~5.0.1",
-    "tsickle": "^0.34.0",
-    "tslib": "^1.9.0",
+    "tsickle": "^0.36.0",
+    "tslib": "^1.10.0",
     "tslint": "~5.9.1",
-    "typescript": "~3.1"
+    "typescript": "~3.5.3"
   }
 }
